Modified file structure, namespaces and modeling
Now all the models are under the TelegramBot::Model namespace. Also, all
the attachments of messages are in the
TelegramBot::Model::MessageAttachment namespace.

Now we use ActiveModel to model our data. Also, there is a conversion
code on TelegramBot::Model::Base class to convert the data received into
it's complex types based on a configuration hash called
FIELD_CONVERSIONS.